// src/generated/models/Ai_alhasbamilestonesesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Ai_alhasbamilestonesesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Ai_alhasbamilestonesesstatecode = keyof typeof Ai_alhasbamilestonesesstatecode;
export const Ai_alhasbamilestonesesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Ai_alhasbamilestonesesstatuscode = keyof typeof Ai_alhasbamilestonesesstatuscode;

export interface Ai_alhasbamilestonesesBase {
  ai_actual_date?: string;
  ai_alhasbamilestonesid: string;
  ai_id?: string;
  ai_name: string;
  ai_planned_date?: string;
  ai_status?: string;
  ai_use_case_id?: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Ai_alhasbamilestonesesstatecode;
  statuscode?: Ai_alhasbamilestonesesstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Ai_alhasbamilestoneses extends Ai_alhasbamilestonesesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}